requirements.txt
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pandas>=2.2.0
openpyxl>=3.1.2
reportlab>=4.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
